LayoutTests/platform/glib/http/tests/websocket/tests/hybi/close-code-and-reason-expected.txt

Test CloseEvent code and reason property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

ws.onclose() was called.
PASS closeEvent.wasClean is false
PASS closeEvent.code is codeAbnormalClosure
PASS closeEvent.reason is emptyString
WebSocketTest.onopen() was called with testId = 0.
WebSocketTest.onclose() was called with testId = 0.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1005
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1005
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 1.
WebSocketTest.onclose() was called with testId = 1.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 2.
WebSocketTest.onclose() was called with testId = 2.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1000
PASS closeEvent.reason is 'ok'
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1000
PASS closeEvent.reason is 'ok'
WebSocketTest.onopen() was called with testId = 3.
WebSocketTest.onclose() was called with testId = 3.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 4.
WebSocketTest.onclose() was called with testId = 4.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 5.
WebSocketTest.onclose() was called with testId = 5.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 6.
WebSocketTest.onclose() was called with testId = 6.
PASS typeof closeEvent.wasClean is 'boolean'
FAIL closeEvent.wasClean should be true. Was false.
FAIL closeEvent.code should be 65535. Was 1006.
FAIL closeEvent.reason should be good bye. Was .
PASS typeof closeEvent.wasClean is 'boolean'
FAIL closeEvent.wasClean should be true. Was false.
FAIL closeEvent.code should be 65535. Was 1006.
FAIL closeEvent.reason should be good bye. Was .
PASS successfullyParsed is true

TEST COMPLETE
